requirements.txt
python-decouple==3.8
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0